+++ application/src/main/java/nl/jixxed/eliteodysseymaterials/domain/ships/core_internals/LifeSupport.java
@@ -96,7 +96,7 @@ public class LifeSupport extends CoreModule {
             LIFE_SUPPORT_8_A
     );
 
-    private LifeSupport(final String id, final HorizonsBlueprintName name, final ModuleSize moduleSize, final ModuleClass moduleClass, final int basePrice, final String internalName, final Map<HorizonsModifier, Object> attributes) {
+    private LifeSupport(final String id, final HorizonsBlueprintName name, final ModuleSize moduleSize, final ModuleClass moduleClass, final long basePrice, final String internalName, final Map<HorizonsModifier, Object> attributes) {
         super(id, name, moduleSize, moduleClass, basePrice, internalName, attributes);
     }
