+++ application/src/main/java/nl/jixxed/eliteodysseymaterials/domain/ships/hardpoint/RemoteReleaseFlechetteLauncher.java
@@ -16,11 +16,11 @@ public class RemoteReleaseFlechetteLauncher extends HardpointModule {
             REMOTE_RELEASE_FLECHETTE_LAUNCHER_2_B_F,
             REMOTE_RELEASE_FLECHETTE_LAUNCHER_2_B_T
     );
-    public RemoteReleaseFlechetteLauncher(String id, HorizonsBlueprintName name, ModuleSize moduleSize, ModuleClass moduleClass, boolean multiCrew, Mounting mounting, int basePrice, String internalName, Map<HorizonsModifier, Object> attributes) {
+    public RemoteReleaseFlechetteLauncher(String id, HorizonsBlueprintName name, ModuleSize moduleSize, ModuleClass moduleClass, boolean multiCrew, Mounting mounting, long basePrice, String internalName, Map<HorizonsModifier, Object> attributes) {
         super(id, name, moduleSize, moduleClass, multiCrew, mounting, basePrice, internalName, attributes);
     }
 
-    public RemoteReleaseFlechetteLauncher(String id, HorizonsBlueprintName name, ModuleSize moduleSize, ModuleClass moduleClass, Origin origin, boolean multiCrew, Mounting mounting, int basePrice, String internalName, Map<HorizonsModifier, Object> attributes) {
+    public RemoteReleaseFlechetteLauncher(String id, HorizonsBlueprintName name, ModuleSize moduleSize, ModuleClass moduleClass, Origin origin, boolean multiCrew, Mounting mounting, long basePrice, String internalName, Map<HorizonsModifier, Object> attributes) {
         super(id, name, moduleSize, moduleClass, origin, multiCrew, mounting, basePrice, internalName, attributes);
     }
