+++ application/src/main/java/nl/jixxed/eliteodysseymaterials/domain/ships/hardpoint/SubSurfaceDisplacementMissile.java
@@ -22,11 +22,11 @@ public class SubSurfaceDisplacementMissile extends HardpointModule {
             SUB_SURFACE_DISPLACEMENT_MISSILE_2_B_T,
             SUB_SURFACE_EXTRACTION_MISSILE_2_B_F
     );
-    public SubSurfaceDisplacementMissile(String id, HorizonsBlueprintName name, ModuleSize moduleSize, ModuleClass moduleClass, boolean multiCrew, Mounting mounting, int basePrice, String internalName, Map<HorizonsModifier, Object> attributes) {
+    public SubSurfaceDisplacementMissile(String id, HorizonsBlueprintName name, ModuleSize moduleSize, ModuleClass moduleClass, boolean multiCrew, Mounting mounting, long basePrice, String internalName, Map<HorizonsModifier, Object> attributes) {
         super(id, name, moduleSize, moduleClass, multiCrew, mounting, basePrice, internalName, attributes);
     }
 
-    public SubSurfaceDisplacementMissile(String id, HorizonsBlueprintName name, ModuleSize moduleSize, ModuleClass moduleClass, Origin origin, boolean multiCrew, Mounting mounting, int basePrice, String internalName, Map<HorizonsModifier, Object> attributes) {
+    public SubSurfaceDisplacementMissile(String id, HorizonsBlueprintName name, ModuleSize moduleSize, ModuleClass moduleClass, Origin origin, boolean multiCrew, Mounting mounting, long basePrice, String internalName, Map<HorizonsModifier, Object> attributes) {
         super(id, name, moduleSize, moduleClass, origin, multiCrew, mounting, basePrice, internalName, attributes);
     }
