+++ application/src/main/java/nl/jixxed/eliteodysseymaterials/domain/ships/optional_internals/CargoRack.java
@@ -32,7 +32,7 @@ public class CargoRack extends OptionalModule {
              CARGO_RACK_8_E
     );
 
-    public CargoRack(String id, HorizonsBlueprintName name, ModuleSize moduleSize, ModuleClass moduleClass, boolean multiCrew, int basePrice, String internalName, Map<HorizonsModifier, Object> attributes) {
+    public CargoRack(String id, HorizonsBlueprintName name, ModuleSize moduleSize, ModuleClass moduleClass, boolean multiCrew, long basePrice, String internalName, Map<HorizonsModifier, Object> attributes) {
         super(id, name, moduleSize, moduleClass, multiCrew, basePrice, internalName, attributes);
     }
